update : 챌린지 상세 조회, 참여하기, 챌린지 참여 취소 로직 수정
1. 챌린지 상세조회,참여하기 : 공개 챌린지이거나 초대받은 챌린지인 경우 가능하도록 수정
2. 챌린지 참여 취소 : 챌린지 생성자는 취소 불가능하도록 수정

+++ src/main/java/com/likelion/devroutine/participant/service/ParticipationService.java
@@ -5,6 +5,7 @@ import com.likelion.devroutine.challenge.exception.ChallengeNotFoundException;
 import com.likelion.devroutine.challenge.exception.InProgressingChallengeException;
 import com.likelion.devroutine.challenge.exception.InaccessibleChallengeException;
 import com.likelion.devroutine.hashtag.dto.ChallengeHashTagResponse;
+import com.likelion.devroutine.invite.repository.InviteRepository;
 import com.likelion.devroutine.participant.domain.Participation;
 import com.likelion.devroutine.participant.dto.ParticipationChallengeDto;
 import com.likelion.devroutine.participant.dto.ParticipationResponse;
@@ -12,6 +13,7 @@ import com.likelion.devroutine.participant.enumerate.ResponseMessage;
 import com.likelion.devroutine.participant.exception.DuplicatedParticipationException;
 import com.likelion.devroutine.challenge.repository.ChallengeRepository;
 import com.likelion.devroutine.participant.exception.ParticipationNotFoundException;
+import com.likelion.devroutine.participant.exception.RejectCancelException;
 import com.likelion.devroutine.participant.repository.ParticipationRepository;
 import com.likelion.devroutine.user.domain.User;
 import com.likelion.devroutine.user.dto.UserResponse;
@@ -30,11 +32,13 @@ public class ParticipationService {
     private final ParticipationRepository participationRepository;
     private final UserRepository userRepository;
     private final ChallengeRepository challengeRepository;
+    private final InviteRepository inviteRepository;
 
-    public ParticipationService(ParticipationRepository participationRepository, UserRepository userRepository, ChallengeRepository challengeRepository) {
+    public ParticipationService(ParticipationRepository participationRepository, UserRepository userRepository, ChallengeRepository challengeRepository, InviteRepository inviteRepository) {
         this.participationRepository = participationRepository;
         this.userRepository = userRepository;
         this.challengeRepository = challengeRepository;
+        this.inviteRepository = inviteRepository;
     }
     @Transactional
     public ParticipationResponse participateChallenge(String oauthId, Long challengeId) {
@@ -52,6 +56,9 @@ public class ParticipationService {
     public ParticipationResponse cancelChallenge(String oauthId, Long challengeId) {
         User user=getUser(oauthId);
         Challenge challenge=getChallenge(challengeId);
+        if(challenge.getUserId().equals(user.getId())){
+            throw new RejectCancelException();
+        }
         Participation participation =getParticipant(user, challenge);
         isProgressChallenge(challenge.getStartDate());
         participationRepository.delete(participation);
@@ -67,6 +74,7 @@ public class ParticipationService {
         List<Participation> participations = participationRepository.findAllByChallenge(challenge);
         return ParticipationChallengeDto.toResponse(participation, ChallengeHashTagResponse.of(challenge.getChallengeHashTags()), UserResponse.toList(getChallengeParticipants(participations)));
     }
+
     public List<User> getChallengeParticipants(List<Participation> participations){
         return participations.stream().map(participant -> participant.getUser())
                 .collect(Collectors.toList());
@@ -97,15 +105,16 @@ public class ParticipationService {
     }
 
     public boolean validateParticipate(User user, Challenge challenge){
-        isVigibility(challenge);
+        isViewable(challenge, user);
         isProgressChallenge(challenge.getStartDate());
         validateDuplicateParticipate(user, challenge);
         return true;
     }
-
-    public boolean isVigibility(Challenge challenge) {
-        if (challenge.getVigibility()) return true;
-        else throw new InaccessibleChallengeException();
+    public boolean isViewable(Challenge challenge, User user){
+        if(challenge.getVigibility() || !inviteRepository.findAllByChallengeIdAndInviteeId(challenge.getId(), user.getId()).isEmpty()){
+            return true;
+        }
+        throw new InaccessibleChallengeException();
     }
 
     public boolean isProgressChallenge(LocalDate startDate) {
